finish main

+++ app/app_common.c
@@ -10,11 +10,11 @@
  * 
  * @return 当前时间戳-毫秒级
  */
-long get_current_time() 
+long long get_current_time() 
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    return (long long)tv.tv_sec * 1000 + tv.tv_usec / 1000;
 }
 
 /**
